0.1.1 AD5660

+++ libraries/AD5660/AD5660.cpp
@@ -1,8 +1,8 @@
 //
 //    FILE: AD5660.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
-//    DATE: 2024-10-25
+// VERSION: 0.1.1
+//    DATE: 2024-10-29
 // PURPOSE: Arduino library for AD5660 Digital Analog Convertor (16 bit).
 
 
@@ -51,7 +51,7 @@ void AD5660::begin()
     //  _mySPI->begin();
     //  delay(1);
   }
-  else  //  SOFTWARE SPI
+  else  //  SOFTWARE SPI MODE0
   {
     pinMode(_dataOut, OUTPUT);
     pinMode(_clock, OUTPUT);
@@ -187,7 +187,7 @@ void AD5660::swSPI_transfer(uint32_t value)
   uint8_t clk = _clock;
   uint8_t dao = _dataOut;
   //  24 bit
-  for (uint16_t mask = 0x800000; mask; mask >>= 1)
+  for (uint32_t mask = 0x800000; mask; mask >>= 1)
   {
     digitalWrite(dao,(value & mask));
     digitalWrite(clk, HIGH);
